Added: Default implementation of the compatibility layer
The default implementation should run on every POSIX compatible shell, but I
only tested it with the busybox ash.

I did not implement every function provided in the compatibility layer of ZSH
and Bash, but only the ones actually used in the colorize code, since that makes
the implementation much simpler.